src/requirements.txt
numpy~=1.26.2
pydub~=0.25.1
Flask~=3.1.0
flask_session~=0.8.0
requests~=2.32.3
faster-whisper~=1.1.1
gunicorn~=23.0.0
Jinja2~=3.1.5
bcrypt~=4.2.1
PyJWT~=2.10.1
colorama~=0.4.6
python-magic~=0.4.27
icad_tone_detection~=2.2
silero-vad~=5.1.2